BASESTATION/requirements.txt
torch==1.9.0
torchvision==0.10.0
Pillow==9.3.0
requests==2.28.1
djitellopy==2.4.0